requirements.txt
asgiref==3.9.1
Django==5.2.4
graphene==3.4.3
graphene-django==3.2.3
graphene-file-upload==1.3.0
graphql-core==3.2.6
graphql-relay==3.2.0
promise==2.3
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.14.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-allauth==0.63.6
django-cors-headers==4.4.0
PyJWT==2.8.0
requests==2.32.3
python-decouple==3.8
django-graphql-jwt==0.4.0
psycopg2-binary==2.9.9
cryptography==45.0.5
setuptools==80.9.0
dj-database-url==2.1.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow==10.0.0
phonenumbers==9.0.10
python-magic==0.4.27
celery==5.5.3
